feat: add new expect api
