+++ expect/_utils.ts
@@ -1,0 +1,7 @@
+export function isPromiseLike(value: unknown): value is PromiseLike<unknown> {
+  if (value == null) {
+    return false;
+  } else {
+    return typeof ((value as Record<string, unknown>).then) === "function";
+  }
+}
